# packages/spec/openapi.yaml
openapi: 3.1.0
info:
  title: Pathfinder API
  description: VEuPathDB Strategy Builder Chatbot API
  version: 1.0.0
paths:
...
  /api/v1/strategies:
    get:
      tags:
      - strategies
      summary: List Strategies
      description: List user's saved strategies.
      operationId: list_strategies_api_v1_strategies_get
      security:
      - APIKeyCookie: []
      parameters:
      - name: siteId
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Siteid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StrategySummaryResponse'
                title: Response List Strategies Api V1 Strategies Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - strategies
      summary: Create Strategy
      description: Create a new strategy.
      operationId: create_strategy_api_v1_strategies_post
      security:
      - APIKeyCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStrategyRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StrategyResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/strategies/{strategyId}:
    get:
      tags:
      - strategies
      summary: Get Strategy
      description: Get a strategy by ID.
      operationId: get_strategy_api_v1_strategies__strategyId__get
      parameters:
      - name: strategyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Strategyid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StrategyResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - strategies
      summary: Update Strategy
      description: Update a strategy.
      operationId: update_strategy_api_v1_strategies__strategyId__patch
      parameters:
      - name: strategyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Strategyid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStrategyRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StrategyResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - strategies
      summary: Delete Strategy
      description: Delete a strategy.
      operationId: delete_strategy_api_v1_strategies__strategyId__delete
      parameters:
      - name: strategyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Strategyid
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/strategies/open:
    post:
      tags:
      - strategies
      summary: Open Strategy
      description: Open a strategy by local or WDK strategy.
      operationId: open_strategy_api_v1_strategies_open_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenStrategyRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenStrategyResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - APIKeyCookie: []
  /api/v1/strategies/wdk:
    get:
      tags:
      - strategies
      summary: List Wdk Strategies
      description: List strategies from VEuPathDB WDK.
      operationId: list_wdk_strategies_api_v1_strategies_wdk_get
      parameters:
      - name: siteId
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Siteid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WdkStrategySummaryResponse'
                title: Response List Wdk Strategies Api V1 Strategies Wdk Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/strategies/wdk/{wdkStrategyId}:
    delete:
      tags:
      - strategies
      summary: Delete Wdk Strategy
      description: Delete a strategy from VEuPathDB WDK.
      operationId: delete_wdk_strategy_api_v1_strategies_wdk__wdkStrategyId__delete
      parameters:
      - name: wdkStrategyId
        in: path
        required: true
        schema:
          type: integer
          title: Wdkstrategyid
      - name: siteId
        in: query
        required: true
        schema:
          type: string
          title: Siteid
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/strategies/wdk/{wdkStrategyId}/import:
    post:
      tags:
      - strategies
      summary: Import Wdk Strategy
      description: Import a WDK strategy as a local snapshot.
      operationId: import_wdk_strategy_api_v1_strategies_wdk__wdkStrategyId__import_post
      security:
      - APIKeyCookie: []
      parameters:
      - name: wdkStrategyId
        in: path
        required: true
        schema:
          type: integer
          title: Wdkstrategyid
      - name: siteId
        in: query
        required: true
        schema:
          type: string
          title: Siteid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StrategyResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/strategies/{strategyId}/push:
    post:
      tags:
      - strategies
      summary: Push To Wdk
      description: Push strategy to VEuPathDB WDK.
      operationId: push_to_wdk_api_v1_strategies__strategyId__push_post
      parameters:
      - name: strategyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Strategyid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PushResultResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/strategies/{strategyId}/sync-wdk:
    post:
      tags:
      - strategies
      summary: Sync Strategy From Wdk
      description: Sync local strategy snapshot from VEuPathDB WDK.
      operationId: sync_strategy_from_wdk_api_v1_strategies__strategyId__sync_wdk_post
      parameters:
      - name: strategyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Strategyid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StrategyResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CreateStrategyRequest:
      properties:
        name:
          type: string
          maxLength: 255
          minLength: 1
          title: Name
        siteId:
          type: string
          title: Siteid
        plan:
          $ref: '#/components/schemas/StrategyPlan-Input'
      type: object
      required:
      - name
      - siteId
      - plan
      title: CreateStrategyRequest
      description: Request to create a strategy.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    OpenStrategyRequest:
      properties:
        strategyId:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Strategyid
        wdkStrategyId:
          anyOf:
          - type: integer
          - type: 'null'
          title: Wdkstrategyid
        siteId:
          anyOf:
          - type: string
          - type: 'null'
          title: Siteid
      type: object
      title: OpenStrategyRequest
      description: Request to open a strategy.
    OpenStrategyResponse:
      properties:
        strategyId:
          type: string
          format: uuid
          title: Strategyid
      type: object
      required:
      - strategyId
      title: OpenStrategyResponse
      description: Open strategy response.
...
    StrategySummaryResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        siteId:
          type: string
          title: Siteid
        recordType:
          anyOf:
          - type: string
          - type: 'null'
          title: Recordtype
        stepCount:
          type: integer
          title: Stepcount
        resultCount:
          anyOf:
          - type: integer
          - type: 'null'
          title: Resultcount
        wdkStrategyId:
          anyOf:
          - type: integer
          - type: 'null'
          title: Wdkstrategyid
        createdAt:
          type: string
          format: date-time
          title: Createdat
        updatedAt:
          type: string
          format: date-time
          title: Updatedat
      type: object
      required:
      - id
      - name
      - siteId
      - recordType
      - stepCount
      - createdAt
      - updatedAt
      title: StrategySummaryResponse
      description: Strategy summary for list views.
...
    UpdateStrategyRequest:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 255
            minLength: 1
          - type: 'null'
          title: Name
        plan:
          anyOf:
          - $ref: '#/components/schemas/StrategyPlan-Input'
          - type: 'null'
        wdkStrategyId:
          anyOf:
          - type: integer
          - type: 'null'
          title: Wdkstrategyid
      type: object
      title: UpdateStrategyRequest
      description: Request to update a strategy.
...
  securitySchemes:
    APIKeyCookie:
      type: apiKey
      in: cookie
      name: pathfinder-auth
servers:
- url: http://localhost:8000
  description: Local development
- url: https://api.pathfinder.veupathdb.org
  description: Production